requirements.txt
ultralytics==8.0.196
opencv-python
torch==2.1.0+cu118
torchvision==0.16.0+cu118
torchaudio==2.1.0+cu118
pandas
numpy
matplotlib
tqdm
Pillow
PyYAML